// frontends/epro/src/app/api/model/workflow-status-dto.ts
/**
 * KV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: DEV-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorkflowDTO } from './workflow-dto';
import { WorkflowStateDTO } from './workflow-state-dto';


export interface WorkflowStatusDTO { 
    pk: number;
    /**
     * The workflow ID
     */
    workflowId: string;
    /**
     * The workflow status ID
     */
    statusId: string;
    /**
     * User message that was attached to the latest status change
     */
    triggerMessage: string;
    workflow: WorkflowDTO;
    state: WorkflowStateDTO;
    /**
     * Date of the WorkflowStatus
     */
    date: Date;
    orderBy?: number;
    /**
     * Does this WorkflowStatus have a creation action ?
     */
    hasCreationAction: boolean;
    /**
     * Scope reference of the WorkflowStatus
     */
    scopeFk: number;
    /**
     * Code of the scope to which WorkflowStatus is attached to
     */
    scopeCode: string;
    /**
     * Shortname of the scope to which WorkflowStatus is attached to
     */
    scopeShortname: string;
    /**
     * Event reference of the WorkflowStatus
     */
    eventFk?: number;
    /**
     * Shortname of the event to which WorkflowStatus is attached to
     */
    eventShortname?: string;
    /**
     * Date of the event to which WorkflowStatus is attached to
     */
    eventDate?: Date;
    /**
     * Field reference of the WorkflowStatus
     */
    fieldFk?: number;
    /**
     * Shortname of the field to which WorkflowStatus is attached to
     */
    fieldShortname?: string;
}